+++ engine/engine.h
@@ -34,7 +34,7 @@ namespace Engine {
 
     Eigen::Matrix3d rotateDcelAlreadyScaled(Dcel& d, unsigned int rot);
 
-    Eigen::Matrix3d scaleAndRotateDcel(Dcel& d, unsigned int rot = 0, int factor = 1);
+    Eigen::Matrix3d scaleAndRotateDcel(Dcel& d, unsigned int rot = 0, double factor = 1);
 
     void getFlippedFaces(std::set<const Dcel::Face*>& flippedFaces, std::set<const Dcel::Face*>& savedFaces, const Dcel& d, const Vec3& target, double angleThreshold, double areaThreshold);
